weston: rename bbappend to adhere to compatible versions only
The patches included in the weston bbappend are only compatible with 1.x
versions of weston. Rename the file so that it only applies to such versions.

NOTE: how these patches should be ported (if at all) to modern versions of
weston is out of the scope of this commit. Additional patching could be needed
when building weston >= 2.0 with drm backend.

Signed-off-by: Gerard Salvatella <gerard.salvatella@toradex.com>