meta: drop useless EXTRA_IMAGEDEPENDS
For the recipes listed in EXTRA_IMAGEDEPENDS, bitbake would create a
dependency as:
| image_pn:do_image_complete -> imagedepends_pn:do_populate_sysroot

but we actually need a image task depend on the deploy tasks of
EXTRA_IMAGEDEPENDS recipes, setting EXTRA_IMAGEDEPENDS would not help
with that, so drop it, use WKS_FILE_DEPENDS instead.

We also need drop "IMAGE_BOOT_FILES +=", that would override
"WKS_FILE_DEPENDS ?=".

Signed-off-by: Ming Liu <ming.liu@toradex.com>